+++ kwcoco/demo/toydata_image.py
@@ -18,7 +18,7 @@ import kwimage
 from kwcoco.demo.toypatterns import CategoryPatterns
 
 try:
-    from xdev import profile
+    from line_profiler import profile
 except Exception:
     profile = ub.identity