+++ kwcoco/metrics/sklearn_alts.py
@@ -9,7 +9,7 @@ import ubelt as ub
 
 
 try:
-    from xdev import profile
+    from line_profiler import profile
 except Exception:
     profile = ub.identity